requirements_CUDA.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.7.0+cu126
torchvision==0.22.0+cu126
diffusers>=0.25.0
transformers>=4.37.0
accelerate>=0.27.0
numpy>=1.26.0
pillow>=10.0.0
matplotlib>=3.8.0
tqdm>=4.66.0
peft
wandb